+++ src/main/java/com/snlabs/aarogyatelangana/account/dao/PatientDao.java
@@ -13,7 +13,7 @@ public interface PatientDao {
     public Patient findByPatientName(String patient);
     public boolean delete(Patient patient);
     public Patient findByPatientId(Patient patient);
-    public Patient searchPatientById(int patientId, UserDetails userDetails);
+    public Patient searchPatientById(long patientId, UserDetails userDetails);
     public Patient searchPatientByName(String patientName, UserDetails userDetails);
     public List<Patient> searchPatientProfilesByCreator(UserDetails userDetails, Form form);
     public List<Patient> listPatientProfilesByDate(Date fromDate, Date toDate, UserDetails userDetails);
